cucumber/requirements.txt
allure-behave==2.5.3
allure-pytest==2.5.3
allure-python-commons==2.5.3
atomicwrites==1.2.1
attrs==18.2.0
behave==1.2.6
certifi==2018.10.15
chardet==3.0.4
idna==2.7
kafka-python==1.4.3
more-itertools==4.3.0
nose==1.3.7
onetimepass==1.0.1
parse==1.9.0
parse-type==0.4.2
pathlib==1.0.1
pathlib2==2.3.2
pluggy==0.8.0
py==1.7.0
pycrypto==2.6.1
pytest==3.10.0
requests==2.20.0
selenium==3.141.0
six==1.11.0
slackclient==1.3.0
urllib3==1.24.1
websocket-client==0.54.0
